Output changes
Initial reservations are now marked as VIP (int:2)
Error message added at the group reservation step, although this should never be needed due to a prior check when finding seat space.
Added message saying "Not Available" if no seats can be found for a group.
Added message to print the remaining number of seats at the end of the program.